implements reservation“

+++ src/main/java/com/modelink/usercenter/service/ChannelService.java
@@ -14,7 +14,7 @@ public interface ChannelService {
      * @param appKey
      * @return
      */
-    public Channel findByAppKey(int appKey);
+    public Channel findByAppKey(Long appKey);
 
     /**
      * 获取渠道列表
